load airport names from apt.dat file
The airport list so far did not show the airport names as they are not contained in the scenery directories.
This commit first parses $FG_ROOT/Airports/apt.dat.gz to build a mapping between the airport ICAO codes and the airport names before loading the remaining data from the scenery folders.
Airport loading time increases significantly, but the list now shows the airport names.